pyproject.toml
[tool.poetry]
name = "gpt-fn"
version = "0.0.0"
description = ""
authors = ["lucemia <lucemia@gmail.com>"]
readme = "README.md"
packages = [{ include = "gpt_fn", from = "src" }]
include = ["fn/py.typed"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27"
pydantic = "^1.0"
jinja2 = "^3.0"

[tool.poetry.group.test.dependencies]
syrupy = "^4.0.2"
pytest-vcr = "^1.0.2"
pytest-cov = "^4.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[tool.poetry-dynamic-versioning]
enable = true
pattern = "default-unprefixed"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
